+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- `-o` now creates the output directory if it does not exist.
+
 ## v0.1.1 - 2023-02-25
 ### Fixed
 - Fix corner cases with text in the summary getting merged.